feat: crud sample

+++ src/admin/pages/DeletePage.js
@@ -1,41 +1,43 @@
 import React, { useContext, useEffect, useState } from "react";
-import { useParams, useNavigate } from "react-router-dom";
 import { SettingsContext } from "../../context/SettingsContext";
+import SelectPage from "../components/SelectPage";
 
 export default function DeletePage() {
-  const { lang } = useContext(SettingsContext);
-  const { id } = useParams();
-  const navigate = useNavigate();
-  const [page, setPage] = useState(null);
+  const { lang, theme } = useContext(SettingsContext);
+  const [pageId, setPageId] = useState(null);
 
-  useEffect(() => {
-    fetch(`/wordpress/wp-json/the-box-admin/v1/pages/${id}`)
-      .then((response) => response.json())
-      .then((data) => setPage(data))
-      .catch((error) => console.error("Error fetching page:", error));
-  }, [id]);
+  function updatePageId(e) {
+    e.preventDefault();
+    const id = e.target.value;
+    setPageId(id);
+  }
 
-  const handleDelete = () => {
-    fetch(`/wordpress/wp-json/the-box-admin/v1/pages/${id}`, {
+  const handleDelete = (e) => {
+    e.preventDefault();
+    if (!pageId) {
+      alert("Please select a page to delete.");
+      return;
+    }
+    fetch(`http://localhost/wordpress/wp-json/the-box-admin/v1/pages/${pageId}`, {
       method: "DELETE",
     })
       .then((response) => response.json())
       .then((data) => {
         alert("Page deleted successfully!");
-        navigate("/list-pages");
       })
       .catch((error) => console.error("Error deleting page:", error));
   };
 
-  if (!page) return <div>Loading...</div>;
 
   return (
-    <div className="the_box-delete-page">
-      <h1>{lang["delete_page"]}</h1>
+    <div className={`the_box-create-page ${theme === 'dark' ? 'dark' : ''}`}>
+    <h1 className="the_box-create-page__title">{lang["delete_page"] || 'Delete Page'}</h1>
+    <h1>{lang["delete_page"]}</h1>
+    <SelectPage onPageSelect={updatePageId} />
       <p>
-        Are you sure you want to delete <strong>{page.title}</strong>?
+        Are you sure you want to delete <strong>{}</strong>?
       </p>
       <button onClick={handleDelete}>Delete</button>
-    </div>
+  </div>
   );
 }
